Fixed too many 'IServiceProvider' instances created issue
I got an exception for 'ManyServiceProvidersCreatedWarning'
when calling any API endpoint more than 20 times.

This was do to how the DbContext was set up using DI. An instance of the
DbContext is created for each request and passed to the controller to
perform a unit-of-work before being disposed when the request ends.

https://learn.microsoft.com/en-us/ef/core/dbcontext-configuration/

Got the initial solution from:
https://stackoverflow.com/a/76697983