Add repository layer
Add two repository classes for Claims and Covers and move the db logic
here, including updating the audit table for create and delete.

This makes the controllers only dependent on the repository class and
moves more logic out of this layer.

Later we can introduce repository interfaces for further abstraction.