Task 2: Add validation when adding a new Cover
Do a check for valid start and end dates when adding a new cover.

The rules are:
 - StartDate cannot be in the past
 - total insurance period cannot exceed 1 year

In addition we check that EndDate is after StartDate

By default, when throwing an exception, ASP.NET will return HTTP status
code 500 - Internal Error. It is more correct to return a 400 - Bad Request
in this case and therefore we use a try catch in the controller and
return an appropriate error code based on the exception.
We use ValidationException here since that is fitting.

According to some, the correct status code would be
422 - Unprocessable Entity. Ref:
https://www.bennadel.com/blog/2434-http-status-codes-for-invalid-data-400-vs-422.htm