Task 4: Adding more tests
Adding tests for all the API endpoints in the ClaimsController.
Part of this was to move the code for setting up the application
into the constructor so that it can be used by all tests.

Notice that the Delete_Claim() test fails. We would expect trying to
delete an Claim with non-existing ID would fail, but current code returns
success. Will fix this in the next commit.