Add error handling in ClaimsController
We did not check if a Claim with the given Id existed in the database
when getting or deleting a Claim. It's better to handle this and give the
caller a proper message and correct HTTP status code.

This also fixed a test, but also showed that one test was wrong -
getting a Claim with a bad Id should not return success.

Throwing exception when a Claim is not found made it not necessary
to return a nullable Claim, this was changed in all the signatures.