Make the ComputePremium method more readable
It's hard to follow the logic with many if-statements and conditions,
so to make it more readable, now the if statements only deals with
the insurance length (number of days).
Then add a separate logic for the different discounts based on vessel types.

Also use the number as given in the requirements to easier see any errors.

The unit test are now helpful to ensure no bugs are introduced in the refactoring.