Add CoversControllerTests
We should keep the logic coherent in our code base,
so add tests for the CoversController similar to the ClaimsController.

This also results in changes to the status code when a give Cover is not
found. Like for claims we return NotFound() when the Id is not found.

Throwing exception when a Cover is not found made it not necessary
to return a nullable Cover, so this was changed in all the signatures.