Avoid auditing blocking HTTP request
Task 3 was about not letting saving the audit records to the SQL database
block the HTTP request, and suggested using some asynchronous patterns.
Since I haven't moved the app to Azure yet, I do it in memory now.

Adding a queue in the Auditer class so that when adding an audit record
from the HTTP request, it is only added to this in-memory queue. The there
is a timer that every second will check for items in the queue and save
them to the db. In case there are a lot of audit items to save, the timer
stopped when saving and restarted after wards.

In order to accomplish this, the auditer class must be added as a singleton
as this now has a queue that we want to live forever. So the classes that
previously created instances of the Auditer class now get's it via DI.

And the methods to add audit records that I previously made async is no
reverted back to synchronous.